requirements.txt
pandas
openpyxl
pyyaml
python-dotenv
openai
google-generativeai
matplotlib
tqdm
